+++ src/anchors/mod.rs
@@ -1,5 +1,1 @@
 pub mod anchor;
-
-use super::eth;
-use super::extensions;
-use super::relay;
