--- src/eth/contracts/mod.rs
@@ -1,5 +1,0 @@
-pub mod queries;
-pub mod signatures;
-
-pub use self::queries::*;
-pub use self::signatures::*;
