+++ src/eth/mod.rs
@@ -1,6 +1,3 @@
 pub mod contracts;
 pub mod transaction;
 pub mod utils;
-
-use super::errors;
-use super::relay;
