+++ src/main.rs
@@ -33,6 +33,7 @@ pub mod anchors;
 pub mod errors;
 pub mod eth;
 pub mod extensions;
+pub mod flush;
 pub mod relay;
 pub mod relay_config;
 pub mod server;
