+++ src/relay.rs
@@ -18,9 +18,8 @@ use super::errors::OperationError;
 use super::eth::contracts::{FLUSH_EVENT_SIGNATURE, TRANSFER_EVENT_SIGNATURE};
 use super::eth::utils::clean_0x;
 use super::extensions::removed::{CancelRemoved, ExitOnLogRemoved};
+use super::flush::{CheckForPastFlush, ProcessFlush};
 use super::server::{HandleRequests, RequestType};
-use super::transfers::flush::CheckForPastFlush;
-use super::transfers::flush::ProcessFlush;
 use super::transfers::live::ProcessTransfer;
 use super::transfers::live::WatchLiveLogs;
 use super::transfers::past::RecheckPastTransferLogs;
