+++ src/relay_config/mod.rs
@@ -1,6 +1,3 @@
 pub mod consul_configs;
 pub mod logger;
 pub mod settings;
-
-use super::errors;
-use super::transfers;
