+++ src/server/endpoint.rs
@@ -8,9 +8,9 @@ use web3::futures::prelude::*;
 use web3::futures::sync::mpsc;
 use web3::types::{H256, U256};
 
-use super::errors::EndpointError;
-use super::eth::utils;
-use super::relay::NetworkType;
+use errors::EndpointError;
+use eth::utils;
+use relay::NetworkType;
 
 pub const HOME: &str = "HOME";
 pub const SIDE: &str = "SIDE";
