+++ src/server/mod.rs
@@ -1,9 +1,5 @@
 pub mod endpoint;
 pub mod handler;
 
-use super::errors;
-use super::eth;
-use super::relay;
-
 pub use self::endpoint::RequestType;
 pub use self::handler::HandleRequests;
