--- src/transfers/flush.rs
@@ -1,762 +1,0 @@
-use std::cmp;
-use std::collections::HashMap;
-use web3::contract::Options;
-use web3::futures::future::join_all;
-use web3::futures::future::{ok, Either};
-use web3::futures::prelude::*;
-use web3::futures::sync::mpsc;
-use web3::futures::try_ready;
-use web3::types::{Address, BlockNumber, Bytes, FilterBuilder, Log, TransactionReceipt, H256, U256};
-use web3::DuplexTransport;
-
-use eth::contracts::{
-    FeeQuery, FeeWallet, FeeWalletQuery, Fees, FlushBlock, FlushBlockQuery, FLUSH_EVENT_SIGNATURE,
-    TRANSFER_EVENT_SIGNATURE,
-};
-use eth::transaction::SendTransaction;
-use relay::Network;
-use server::handler::{BalanceOf, BalanceQuery};
-use transfers::live::Event;
-use transfers::transfer::Transfer;
-use transfers::withdrawal::{ApproveParams, WaitForWithdrawalProcessed};
-
-/// Simple struct with an Address and Balance that represents an ethereum wallet
-#[derive(Clone, Copy, Ord, Eq, PartialEq, PartialOrd)]
-pub struct Wallet {
-    address: Address,
-    balance: U256,
-}
-
-impl Wallet {
-    /// Create a new wallet
-    /// # Arguments
-    ///
-    /// * `address` - Address for the wallet
-    /// * `balance` - Current balance of tokens at the wallet
-    fn new(address: &Address, balance: &U256) -> Self {
-        Wallet {
-            address: *address,
-            balance: *balance,
-        }
-    }
-
-    /// Creates a transfer for this wallet given the flush tx hash, block hash, and the block number with offset
-    /// # Arguments
-    ///
-    /// * `transaction_hash`- Transaction hash of the flush
-    /// * `block_hash` - Block hash of the flush
-    /// * `block_number` - Modified block number of the flush. Will not match the actual block number to avoid collisions in the contract
-    fn get_transfer(&self, transaction_hash: &H256, block_hash: &H256, block_number: &U256) -> Transfer {
-        Transfer {
-            destination: self.address,
-            amount: self.balance,
-            tx_hash: *transaction_hash,
-            block_hash: *block_hash,
-            block_number: *block_number,
-            removed: false,
-        }
-    }
-
-    /// Create a SendTransaction Future that performs a withdrawal on the target chain
-    /// # Arguments
-    ///
-    /// * `target` - Network to withdrawl from
-    /// * `transaction_hash`- Transaction hash of the flush
-    /// * `block_hash` - Block hash of the flush
-    /// * `block_number` - Modified block number of the flush. Will not match the actual block number to avoid collisions in the contract
-    fn withdraw<T: DuplexTransport + 'static>(
-        &self,
-        target: &Network<T>,
-        transaction_hash: &H256,
-        block_hash: &H256,
-        block_number: &U256,
-    ) -> SendTransaction<T, ApproveParams> {
-        let approve_params = ApproveParams {
-            destination: self.address,
-            amount: self.balance,
-            tx_hash: *transaction_hash,
-            block_hash: *block_hash,
-            block_number: *block_number,
-        };
-        SendTransaction::new(&target, "approveWithdrawal", &approve_params, target.retries)
-    }
-}
-
-enum CheckForPastFlushState {
-    CheckFlushBlock(Box<Future<Item = FlushBlock, Error = ()>>),
-    GetFlushLog(Box<Future<Item = Vec<Log>, Error = ()>>),
-    GetFlushReceipt(Box<Log>, Box<Future<Item = Option<TransactionReceipt>, Error = ()>>),
-}
-
-/// Future for checking if the flush block is set in relay, then getting the transaction receipt if so
-pub struct CheckForPastFlush<T: DuplexTransport + 'static> {
-    state: CheckForPastFlushState,
-    source: Network<T>,
-}
-
-impl<T: DuplexTransport + 'static> CheckForPastFlush<T> {
-    /// Create a new CheckForPastFlush Future
-    /// # Arguments
-    ///
-    /// * `source` - Network being flushed
-    /// * `tx` - Sender to trigger Flush event processor
-    pub fn new(source: &Network<T>) -> Self {
-        let flush_block_future = source
-            .relay
-            .query::<FlushBlock, Address, BlockNumber, FlushBlockQuery>(
-                "flushBlock",
-                FlushBlockQuery::default(),
-                source.account,
-                Options::default(),
-                BlockNumber::Latest,
-            )
-            .map_err(|e| {
-                error!("error retrieving flush block: {:?}", e);
-            });
-        CheckForPastFlush {
-            state: CheckForPastFlushState::CheckFlushBlock(Box::new(flush_block_future)),
-            source: source.clone(),
-        }
-    }
-    /// Create a Future to get the flush even logs given the block the event occurred at
-    /// # Arguments
-    ///
-    /// * `block` - Block number for the flush event
-    pub fn get_flush_log(&self, block_number: u64) -> Box<Future<Item = Vec<Log>, Error = ()>> {
-        // Not sure if there is a better way to get the log we want, probably get block or something
-        let filter = FilterBuilder::default()
-            .address(vec![self.source.relay.address()])
-            .from_block(BlockNumber::from(block_number - 1))
-            .to_block(BlockNumber::Number(block_number + 1))
-            .topics(Some(vec![FLUSH_EVENT_SIGNATURE.into()]), None, None, None)
-            .build();
-        Box::new(self.source.web3.eth().logs(filter).map_err(|e| {
-            error!("error getting flush log: {:?}", e);
-        }))
-    }
-}
-
-impl<T: DuplexTransport + 'static> Future for CheckForPastFlush<T> {
-    type Item = Option<Event>;
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        loop {
-            let source = self.source.clone();
-            let next = match self.state {
-                CheckForPastFlushState::CheckFlushBlock(ref mut future) => {
-                    let block = try_ready!(future.poll());
-                    let block_number = block.0.as_u64();
-                    info!("flush block on start is {}", block_number);
-                    if block_number > 0 {
-                        let future = self.get_flush_log(block_number);
-                        CheckForPastFlushState::GetFlushLog(future)
-                    } else {
-                        return Ok(Async::Ready(None));
-                    }
-                }
-                CheckForPastFlushState::GetFlushLog(ref mut stream) => {
-                    let logs = try_ready!(stream.poll());
-                    if !logs.is_empty() {
-                        let log = &logs[0];
-                        let removed = log.removed.unwrap_or(false);
-                        match log.transaction_hash {
-                            Some(tx_hash) => {
-                                let future = source.get_receipt(removed, tx_hash);
-                                CheckForPastFlushState::GetFlushReceipt(Box::new(log.clone()), Box::new(future))
-                            }
-                            None => {
-                                error!("flush log missing transaction hash");
-                                return Err(());
-                            }
-                        }
-                    } else {
-                        error!("Didn't find any flush event at flushBlock");
-                        return Ok(Async::Ready(None));
-                    }
-                }
-                CheckForPastFlushState::GetFlushReceipt(ref log, ref mut future) => {
-                    let receipt_option = try_ready!(future.poll());
-                    match receipt_option {
-                        Some(receipt) => {
-                            return Ok(Async::Ready(Some(Event::new(&*log.clone(), &receipt))));
-                        }
-                        None => {
-                            error!("error getting flush receipt");
-                            return Err(());
-                        }
-                    }
-                }
-            };
-            self.state = next;
-        }
-    }
-}
-
-enum ProcessFlushState<T: DuplexTransport + 'static> {
-    Wait,
-    CheckBalances(CheckBalances<T>),
-    FilterContracts(FilterContracts),
-    FilterLowBalance(FilterLowBalance),
-    WithdrawWallets(Box<Future<Item = Vec<()>, Error = ()>>),
-    WithdrawLeftovers(WithdrawLeftovers<T>),
-}
-
-/// Process Flush Log/Receipts that come across and trigger a multi step flush
-/// 1. Check all balances
-/// 1. Filter contracts out
-/// 1. Filter out balances that don't have more than the fee cost to withdraw
-/// 1. Withdraw all balances to the same wallet on target chain
-/// 1. Withdraw any leftovers in the contract after all withdrawals confirmed
-pub struct ProcessFlush<T: DuplexTransport + 'static> {
-    state: ProcessFlushState<T>,
-    source: Network<T>,
-    target: Network<T>,
-    rx: mpsc::UnboundedReceiver<Event>,
-    flush: Option<Event>,
-}
-
-impl<T: DuplexTransport + 'static> ProcessFlush<T> {
-    /// Create a new ProcessFlush Future
-    /// # Arguments
-    ///
-    /// * `source` - Network being flushed
-    /// * `target` - Network to send all the token balances
-    /// * `rx` - Receiver that triggers on Flush event
-    pub fn new(source: &Network<T>, target: &Network<T>, rx: mpsc::UnboundedReceiver<Event>) -> Self {
-        ProcessFlush {
-            state: ProcessFlushState::Wait,
-            source: source.clone(),
-            target: target.clone(),
-            rx,
-            flush: None,
-        }
-    }
-
-    fn handle_flush_event(&self, flush: Event) -> Result<Option<CheckBalances<T>>, ()> {
-        if let Ok(mut lock) = self.source.flushed.write() {
-            *lock = Some(flush.clone());
-        } else {
-            error!("error getting write lock on process");
-            return Err(());
-        }
-
-        info!("flush event triggered");
-        let removed = flush.log.removed.unwrap_or(false);
-        if removed {
-            Ok(None)
-        } else {
-            if flush.receipt.block_hash.is_none() {
-                error!("Failed to get block hash for flush");
-                return Err(());
-            }
-
-            if flush.receipt.block_number.is_none() {
-                error!("Failed to get block hash for flush");
-                return Err(());
-            }
-
-            Ok(Some(CheckBalances::new(&self.source, flush.receipt.block_number)))
-        }
-    }
-
-    fn handle_final_wallets(
-        &self,
-        flush_event: &Event,
-        fees: &U256,
-        wallets: &mut Vec<Wallet>,
-    ) -> Result<impl Future<Item = Vec<()>, Error = ()>, ()> {
-        let receipt = flush_event.receipt.clone();
-        if receipt.block_number.is_none() {
-            error!("no block number in transfer receipt");
-            return Err(());
-        }
-
-        if receipt.block_hash.is_none() {
-            error!("no block hash in transfer receipt");
-            return Err(());
-        }
-
-        let block_hash = receipt.block_hash.unwrap();
-        let block_number = receipt.block_number.unwrap();
-        let fees = *fees;
-        wallets.sort();
-
-        let futures: Vec<Box<Future<Item = (), Error = ()>>> = wallets
-            .iter()
-            .enumerate()
-            .map(|(i, wallet)| {
-                let target = self.target.clone();
-                let wait_target = self.target.clone();
-                let transaction_hash = receipt.transaction_hash;
-                let wallet = *wallet;
-
-                // Create futures
-                let transfer = wallet.get_transfer(&receipt.transaction_hash, &block_hash, &(block_number + i));
-                let future: Box<Future<Item = (), Error = ()>> = Box::new(
-                    transfer
-                        .check_withdrawal(&target, Some(fees))
-                        .and_then(move |needs_approval| {
-                            let target = target.clone();
-                            if needs_approval {
-                                Either::A(wallet.withdraw(&target, &transaction_hash, &block_hash, &(block_number + i)))
-                            } else {
-                                Either::B(ok(()))
-                            }
-                        })
-                        .and_then(move |_| {
-                            let target = wait_target.clone();
-                            // Wait until it is processed before moving on
-                            WaitForWithdrawalProcessed::new(&target, &transfer)
-                        }),
-                );
-                future
-            })
-            .collect();
-        Ok(join_all(futures))
-    }
-}
-
-impl<T: DuplexTransport + 'static> Future for ProcessFlush<T> {
-    type Item = ();
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let target = self.target.clone();
-        let source = self.source.clone();
-        loop {
-            let flush = self.flush.clone();
-            let next: ProcessFlushState<T> = match self.state {
-                ProcessFlushState::Wait => {
-                    let event = try_ready!(self.rx.poll());
-                    match event {
-                        Some(flush) => {
-                            self.flush = Some(flush.clone());
-                            if let Some(future) = self.handle_flush_event(flush)? {
-                                ProcessFlushState::CheckBalances(future)
-                            } else {
-                                ProcessFlushState::Wait
-                            }
-                        }
-                        None => {
-                            return Ok(Async::Ready(()));
-                        }
-                    }
-                }
-                ProcessFlushState::CheckBalances(ref mut future) => {
-                    let balances = try_ready!(future.poll());
-                    info!("found {} wallets with tokens", balances.len());
-                    let contract_future = FilterContracts::new(&source, balances);
-                    ProcessFlushState::FilterContracts(contract_future)
-                }
-                ProcessFlushState::FilterContracts(ref mut future) => {
-                    let balances = try_ready!(future.poll());
-                    info!("{} balances were not contracts", balances.len());
-                    let low_balance_future = FilterLowBalance::new(&target, balances);
-                    ProcessFlushState::FilterLowBalance(low_balance_future)
-                }
-                ProcessFlushState::FilterLowBalance(ref mut future) => {
-                    let (fees, mut balances) = try_ready!(future.poll());
-                    info!("{} wallets above minimum balances", balances.len());
-                    match flush {
-                        Some(flush_event) => {
-                            let futures = self.handle_final_wallets(&flush_event, &fees.0, &mut balances)?;
-                            ProcessFlushState::WithdrawWallets(Box::new(futures))
-                        }
-                        None => {
-                            error!("No flush receipt available");
-                            return Err(());
-                        }
-                    }
-                }
-                ProcessFlushState::WithdrawWallets(ref mut future) => {
-                    let withdrawals = try_ready!(future.poll());
-                    info!("finished {} wallet withdrawals", withdrawals.len());
-                    if let Some(flush_event) = flush {
-                        ProcessFlushState::WithdrawLeftovers(WithdrawLeftovers::new(
-                            &target,
-                            &flush_event.receipt,
-                            withdrawals.len() + 1,
-                        ))
-                    } else {
-                        error!("No flush receipt available");
-                        return Err(());
-                    }
-                }
-                ProcessFlushState::WithdrawLeftovers(ref mut future) => {
-                    try_ready!(future.poll());
-                    info!("finished leftover withdrawal");
-                    return Ok(Async::Ready(()));
-                }
-            };
-            self.state = next;
-        }
-    }
-}
-
-enum WithdrawLeftoversState<T: DuplexTransport + 'static> {
-    GetBalance(Box<Future<Item = BalanceOf, Error = ()>>),
-    GetFeeWallet(Box<Future<Item = FeeWallet, Error = ()>>),
-    Withdraw(SendTransaction<T, ApproveParams>),
-}
-
-/// Future that gets the balance of the relay contract on target chain.
-/// Then, it withdraws that balance to the fee wallet
-pub struct WithdrawLeftovers<T: DuplexTransport + 'static> {
-    state: WithdrawLeftoversState<T>,
-    target: Network<T>,
-    receipt: TransactionReceipt,
-    offset: usize,
-    balance: Option<U256>,
-}
-
-impl<T: DuplexTransport + 'static> WithdrawLeftovers<T> {
-    /// Create a new WithdrawLeftovers Future
-    /// # Arguments
-    ///
-    /// * `source` - Network being flushed
-    /// * `target` - Network to send all the token balances
-    /// * `flush_receipt` - Transaction receipt for the flush event
-    fn new(target: &Network<T>, flush_receipt: &TransactionReceipt, offset: usize) -> Self {
-        let state = WithdrawLeftoversState::GetBalance(WithdrawLeftovers::get_balance(&target));
-        WithdrawLeftovers {
-            state,
-            target: target.clone(),
-            receipt: flush_receipt.clone(),
-            offset,
-            balance: None,
-        }
-    }
-
-    /// Get a future to find the balance of the relay address
-    /// # Arguments
-    ///
-    /// * `target` - Network being flushed
-    fn get_balance(target: &Network<T>) -> Box<Future<Item = BalanceOf, Error = ()>> {
-        let relay_contract_balance_query = BalanceQuery::new(target.relay.address());
-        let target = target.clone();
-        Box::new(
-            target
-                .token
-                .query::<BalanceOf, Address, BlockNumber, BalanceQuery>(
-                    "balanceOf",
-                    relay_contract_balance_query,
-                    target.account,
-                    Options::default(),
-                    BlockNumber::Latest,
-                )
-                .map_err(|e| {
-                    error!("error retrieving contract balance: {:?}", e);
-                }),
-        )
-    }
-
-    /// Get a future to find the fee wallet for the target chain
-    /// # Arguments
-    ///
-    /// * `target` - Network being flushed
-    fn get_fee_wallet(target: &Network<T>) -> Box<Future<Item = FeeWallet, Error = ()>> {
-        let target = target.clone();
-        Box::new(
-            target
-                .relay
-                .query::<FeeWallet, Address, BlockNumber, FeeWalletQuery>(
-                    "feeWallet",
-                    FeeWalletQuery::default(),
-                    target.account,
-                    Options::default(),
-                    BlockNumber::Latest,
-                )
-                .map_err(|e| {
-                    error!("error retrieving fee wallet: {:?}", e);
-                }),
-        )
-    }
-}
-
-impl<T: DuplexTransport + 'static> Future for WithdrawLeftovers<T> {
-    type Item = ();
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        loop {
-            let target = self.target.clone();
-            let receipt = self.receipt.clone();
-            let offset = self.offset;
-            let next = match self.state {
-                WithdrawLeftoversState::GetBalance(ref mut future) => {
-                    let balance = try_ready!(future.poll());
-                    self.balance = Some(balance.0);
-                    WithdrawLeftoversState::GetFeeWallet(WithdrawLeftovers::get_fee_wallet(&target))
-                }
-                WithdrawLeftoversState::GetFeeWallet(ref mut future) => {
-                    let address = try_ready!(future.poll());
-                    match self.balance {
-                        Some(b) => {
-                            if receipt.block_number.is_none() {
-                                error!("no block number in transfer receipt");
-                                return Err(());
-                            }
-
-                            if receipt.block_hash.is_none() {
-                                error!("no block hash in transfer receipt");
-                                return Err(());
-                            }
-                            let block_hash = receipt.block_hash.unwrap();
-                            let block_number = receipt.block_number.unwrap() + offset;
-                            let wallet = Wallet::new(&address.0, &b);
-                            info!("withdrawing {} to fee wallet {}", wallet.balance, wallet.address);
-                            let withdrawal =
-                                wallet.withdraw(&target, &receipt.transaction_hash, &block_hash, &block_number);
-                            WithdrawLeftoversState::Withdraw(withdrawal)
-                        }
-                        None => {
-                            error!("error balance is none");
-                            return Err(());
-                        }
-                    }
-                }
-                WithdrawLeftoversState::Withdraw(ref mut future) => {
-                    try_ready!(future.poll());
-                    return Ok(Async::Ready(()));
-                }
-            };
-            self.state = next;
-        }
-    }
-}
-
-/// FilterLowBalance Future that takes a list of wallets, and filters out that have a balance below the fee cost to withdraw
-pub struct FilterLowBalance {
-    future: Box<Future<Item = Fees, Error = ()>>,
-    wallets: Vec<Wallet>,
-}
-
-impl FilterLowBalance {
-    /// Create a new FilterContracts Future
-    /// # Arguments
-    ///
-    /// * `source` - Network where the wallets may be contracts
-    /// * `wallets` - Vector of tuples with Address and Balances
-    pub fn new<T: DuplexTransport + 'static>(target: &Network<T>, wallets: Vec<Wallet>) -> Self {
-        // Get contract data
-        let future = target
-            .relay
-            .query::<Fees, Address, BlockNumber, FeeQuery>(
-                "fees",
-                FeeQuery::default(),
-                target.account,
-                Options::default(),
-                BlockNumber::Latest,
-            )
-            .map_err(|e| {
-                error!("error getting relay fees: {:?}", e);
-            });
-
-        FilterLowBalance {
-            future: Box::new(future),
-            wallets,
-        }
-    }
-}
-
-impl Future for FilterLowBalance {
-    type Item = (Fees, Vec<Wallet>);
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let fee = try_ready!(self.future.poll());
-        let fee_value = fee.0;
-        let filtered = self
-            .wallets
-            .iter()
-            .filter_map(|wallet| {
-                if wallet.balance > fee_value {
-                    Some(*wallet)
-                } else {
-                    None
-                }
-            })
-            .collect();
-        Ok(Async::Ready((fee, filtered)))
-    }
-}
-
-/// FilterContract Future that takes a list of wallets, and filters out any that are contracts
-pub struct FilterContracts {
-    future: Box<Future<Item = Vec<Bytes>, Error = ()>>,
-    wallets: Vec<Wallet>,
-}
-
-impl FilterContracts {
-    /// Create a new FilterContracts Future
-    /// # Arguments
-    ///
-    /// * `source` - Network where the wallets may be contracts
-    /// * `wallets` - Vector of tuples with Address and Balances
-    pub fn new<T: DuplexTransport + 'static>(source: &Network<T>, wallets: Vec<Wallet>) -> Self {
-        // Get contract data
-        // For whatever reason, doing this as an iter().map().collect() did not work
-        let mut futures = Vec::new();
-        for wallet in wallets.clone() {
-            let future = source.web3.eth().code(wallet.address, None).map_err(move |e| {
-                error!("error retrieving code for wallet {}: {:?}", wallet.address, e);
-            });
-            futures.push(Box::new(future));
-        }
-
-        FilterContracts {
-            future: Box::new(join_all(futures)),
-            wallets: wallets.clone(),
-        }
-    }
-}
-
-impl Future for FilterContracts {
-    type Item = Vec<Wallet>;
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let all_bytes = try_ready!(self.future.poll());
-        let filtered = self
-            .wallets
-            .iter()
-            .zip(all_bytes)
-            .filter_map(
-                move |(wallet, ref bytes)| {
-                    if bytes.0.is_empty() {
-                        Some(*wallet)
-                    } else {
-                        None
-                    }
-                },
-            )
-            .collect();
-        Ok(Async::Ready(filtered))
-    }
-}
-
-pub enum CheckBalancesState {
-    GetEndingBlock(Box<Future<Item = U256, Error = ()>>),
-    GetLogWindow(u64, u64, Box<Future<Item = Vec<Log>, Error = ()>>),
-}
-
-/// Get all balances for all wallets with tokens by looking over the entire history of the chain.
-/// Keep track of all balances, and add and remove as transfers occur
-/// In order to avoid crashes, it does this by looking at windows of 1000 blocks rather than the whole chain
-pub struct CheckBalances<T: DuplexTransport + 'static> {
-    source: Network<T>,
-    state: CheckBalancesState,
-    balances: HashMap<Address, U256>,
-}
-
-impl<T: DuplexTransport + 'static> CheckBalances<T> {
-    /// Create a new CheckBalance Future
-    /// # Arguments
-    ///
-    /// * `source` - Network where the transfers were performed
-    /// * `block` - Optional ending block, gets latest if None
-    fn new(source: &Network<T>, block: Option<U256>) -> Self {
-        let state = match block {
-            Some(b) => {
-                let window_end = cmp::min(b.as_u64(), 1000);
-                CheckBalancesState::GetLogWindow(
-                    b.as_u64(),
-                    window_end,
-                    CheckBalances::build_next_window(source, 0, window_end),
-                )
-            }
-            None => {
-                let future = source.web3.eth().block_number().map_err(move |e| {
-                    error!("error getting block number {:?}", e);
-                });
-                CheckBalancesState::GetEndingBlock(Box::new(future))
-            }
-        };
-        CheckBalances {
-            source: source.clone(),
-            state,
-            balances: HashMap::new(),
-        }
-    }
-
-    /// Build a filter and look at all logs in the given range for transfer events.
-    /// # Arguments
-    ///
-    /// * `source` - Network where the transfers were performed
-    /// * `start` - Start of the next window
-    /// * `end` - End of the next window
-    fn build_next_window(source: &Network<T>, start: u64, end: u64) -> Box<Future<Item = Vec<Log>, Error = ()>> {
-        let token_address: Address = source.token.address();
-        let filter = FilterBuilder::default()
-            .address(vec![token_address])
-            .from_block(BlockNumber::from(start))
-            .to_block(BlockNumber::Number(end))
-            .topics(Some(vec![TRANSFER_EVENT_SIGNATURE.into()]), None, None, None)
-            .build();
-        //self.state = ;
-        let future = source.web3.eth().logs(filter).map_err(move |e| {
-            error!("error getting block number {:?}", e);
-        });
-        Box::new(future)
-    }
-}
-
-impl<T: DuplexTransport + 'static> Future for CheckBalances<T> {
-    type Item = Vec<Wallet>;
-    type Error = ();
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let source = self.source.clone();
-        loop {
-            let next = match self.state {
-                CheckBalancesState::GetEndingBlock(ref mut future) => {
-                    let block = try_ready!(future.poll());
-                    let window_end = cmp::min(block.as_u64(), 1000);
-                    let future = CheckBalances::build_next_window(&source, 0, window_end);
-                    CheckBalancesState::GetLogWindow(block.as_u64(), window_end, future)
-                }
-                CheckBalancesState::GetLogWindow(end, window_end, ref mut future) => {
-                    let logs = try_ready!(future.poll());
-                    debug!(
-                        "found {} logs with transfers between {} and {}",
-                        logs.len(),
-                        window_end - 1000,
-                        window_end,
-                    );
-                    // Process existing logs
-                    logs.iter().for_each(|log| {
-                        if Some(true) != log.removed {
-                            let sender_address: Address = log.topics[1].into();
-                            let receiver_address: Address = log.topics[2].into();
-                            let amount: U256 = log.data.0[..32].into();
-                            debug!("{} transferred {} to {}", sender_address, amount, receiver_address);
-                            // Don't care if source doesn't exist, because it is likely a mint in that case
-                            let zero = U256::zero();
-                            self.balances
-                                .entry(sender_address)
-                                .and_modify(|v| {
-                                    if !v.is_zero() {
-                                        *v -= amount;
-                                    }
-                                })
-                                .or_insert(zero);
-                            let dest_balance = self.balances.entry(receiver_address).or_insert(zero);
-                            *dest_balance += amount;
-                        }
-                    });
-
-                    debug!("Window end is {} of {} blocks", window_end, end);
-                    // Setup next window
-                    if window_end < end {
-                        let next_window_end = cmp::min(end, window_end + 1000);
-                        let future = CheckBalances::build_next_window(&source, window_end + 1, next_window_end);
-                        CheckBalancesState::GetLogWindow(end, next_window_end, future)
-                    } else {
-                        return Ok(Async::Ready(
-                            self.balances
-                                .iter()
-                                .map(|(key, value)| Wallet::new(key, value))
-                                .collect(),
-                        ));
-                    }
-                }
-            };
-            self.state = next;
-        }
-    }
-}
