+++ src/transfers/mod.rs
@@ -1,10 +1,4 @@
-pub mod flush;
 pub mod live;
 pub mod past;
 pub mod transfer;
 pub mod withdrawal;
-
-use super::eth;
-use super::extensions;
-use super::relay;
-use super::transfers;
